requirements.txt
pandas>=2.0.0
scikit-learn>=1.3.0
transformers>=4.30.0
torch>=2.0.0
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
streamlit>=1.28.0
snscrape>=0.7.0
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.0.0
emoji>=2.0.0
joblib>=1.3.0
accelerate>=0.26.0
datasets>=2.0.0
# Testing dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.24.0
# Logging and monitoring
loguru>=0.7.0
